update decimal to float in schema

+++ app/schemas/dashboard.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 import uuid
 from pydantic import BaseModel
 from app.schemas.payment import PaymentStatus
@@ -8,12 +7,12 @@ class Unit(BaseModel):
     name: str
     projectName: str
     status: str
-    price: Decimal
+    price: float
     image: str | None
 
 class Payment(BaseModel):
     id: uuid.UUID
-    amount: Decimal
+    amount: float
     payment_date: str
     status: PaymentStatus
     reason_for_payment: str | None
@@ -25,8 +24,8 @@ class MonthlyRevenueItem(BaseModel):
 
 class DashboardSummary(BaseModel):
     total_units: int
-    total_revenue: Decimal
-    total_outstanding: Decimal
+    total_revenue: float
+    total_outstanding: float
     total_projects: int
     total_users: int
     total_payments: int
